refactor/move children api calls to action folder
