+++ nursery/components/CheckChild.tsx
@@ -1,7 +1,8 @@
-import { ChildrenInfo } from '@/types/types'
+import { ChildrenInfo } from '@/types/children'
 import React from 'react'
 import styles from "./child.module.css";
 import { format } from 'date-fns';
+import { checkInChild, checkOutChild } from '@/actions/children';
 
 
 export default function CheckChild(child: ChildrenInfo) {
@@ -9,43 +10,7 @@ export default function CheckChild(child: ChildrenInfo) {
   const now = format(Date(), "HH:mm")
   const [time, setTime] = React.useState<string>(now)
 
-  const checkInChild = async (childId: string, time: string) => {
-    const accessToken = process.env.NEXT_PUBLIC_ACCESS_TOKEN
 
-    try {
-      let res = await fetch(`https://app.famly.co/api/v2/children/${childId}/checkins`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          accessToken,
-          pickupTime: time
-        },)
-      })
-      let data = await res.json()
-    } catch (error) {
-      console.log(error)
-    }
-  }
-
-  const checkOutChild = async (childId: string) => {
-    const accessToken = process.env.NEXT_PUBLIC_ACCESS_TOKEN
-    try {
-      let res = await fetch(`https://app.famly.co/api/v2/children/${childId}/checkout`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          accessToken,
-        },)
-      })
-      let data = await res.json()
-    } catch (error) {
-      console.log(error)
-    }
-  }
 
   return (
     <div style={{ display: "flex", flexDirection: "column", alignItems: "center" }}>
